Creating an easier to change website
I have built out the cms and associated templates. Now, adding a
department to the menu bar is as easy as creating one in the CMS. To
add sections for that department you simply create sections, give them
an order by entering different numbers, and create a relationship
between them and a department.

I have set it up so that it is mandatory for departments to have an
associated jpeg uploaded to them. This jpeg will be used to allow
department heads to have some customization of their landing pages.
Bonus: If they want to upload a photo for the section page they can do
that as well(though I did not set it as required), allowing even more
granular control of the look of their department’s article lists.

Next I will do more work in the CMS to tie articles to their respective
departments.